ui: Move echo integration test into separate module
